feat: 🎸 conference details
Added conference location and proper images
